fix oss-fuzz 10428.

+++ qemu/target-sparc/translate.c
@@ -2729,7 +2729,7 @@ static void disas_sparc_insn(DisasContext * dc, unsigned int insn, bool hook_ins
         break;
     case 1:                     /*CALL*/
         {
-            target_long target = GET_FIELDs(insn, 2, 31) << 2;
+            target_long target = (int)(((unsigned int)(GET_FIELDs(insn, 2, 31))) << 2);
             TCGv o7 = gen_dest_gpr(dc, 15);
 
             tcg_gen_movi_tl(tcg_ctx, o7, dc->pc);
